refactor: test utils paths
